app/requirements.txt
absl-py==0.8.0
annoy==1.16.0
astor==0.8.0
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
firebase-admin==2.14.0
Flask==1.1.1
Flask-WTF==0.14.2
gast==0.3.2
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==0.28.1
google-cloud-firestore==0.30.0
google-cloud-storage==1.13.0
google-pasta==0.1.7
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.23.0
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
msgpack==0.6.2
numpy==1.17.2
pandas==0.25.1
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
tb-nightly==1.14.0a20190603
tensorflow==2.0.0b1
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
urllib3==1.25.5
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
gunicorn==19.9.0